Add files via upload
Participants take turns counting from 1 onwards, but with a few twists. Instead of saying the actual number, certain numbers are replaced with words based on specific rules.

The rules are as follows:

If the number is divisible by 3, the player says "Fizz" instead of the number.
If the number is divisible by 5, the player says "Buzz" instead of the number.
If the number is divisible by both 3 and 5, the player says "FizzBuzz" instead of the number.

Made by Alexander Eraso.